Move MediaDecoder::DecodedStreamData and related code to its own file
