+++ dom/media/moz.build
@@ -87,6 +87,7 @@ EXPORTS += [
     'AudioStream.h',
     'BufferMediaResource.h',
     'CubebUtils.h',
+    'DecodedStream.h',
     'DecoderTraits.h',
     'DOMMediaStream.h',
     'EncodedBufferCache.h',
@@ -175,6 +176,7 @@ UNIFIED_SOURCES += [
     'AudioTrack.cpp',
     'AudioTrackList.cpp',
     'CubebUtils.cpp',
+    'DecodedStream.cpp',
     'DOMMediaStream.cpp',
     'EncodedBufferCache.cpp',
     'FileBlockCache.cpp',
